quick sort

+++ 3-quick_sort.c
@@ -8,7 +8,7 @@
  */
 void swap(int *x, int *y)
 {
-	size_t temp = *x;
+	int temp = *x;
 	*x = *y;
 	*y = temp;
 }
